requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic[email]
greenlet
